split linter code into linter-integration.js
