+++ lib/build.js
@@ -86,7 +86,7 @@ export default {
     this.statusBarView && this.statusBarView.destroy();
     this.buildView && this.buildView.destroy();
     this.saveConfirmView && this.saveConfirmView.destroy();
-    this.linter && this.linter.dispose();
+    this.linter && this.linter.destroy();
     this.targetManager.destroy();
 
     clearTimeout(this.finishedTimer);
@@ -102,7 +102,7 @@ export default {
     const BuildError = require('./build-error');
     const path = require('./utils').activePath();
     let buildTitle = '';
-    this.linter && this.linter.deleteMessages();
+    this.linter && this.linter.clear();
 
     Promise.resolve(this.targetManager.getTargets(path)).then(targets => {
       if (!targets || 0 === targets.length) {
@@ -187,39 +187,7 @@ export default {
           success = success && !this.errorMatcher.getMatches().some(match => match.type && match.type.toLowerCase() === 'error');
         }
 
-        function extractRange(json) {
-          return [
-            [ (json.line || 1) - 1, (json.col || 1) - 1 ],
-            [ (json.line_end || json.line || 1) - 1, (json.col_end || json.col || 1) - 1 ]
-          ];
-        }
-        function normalizePath(p) {
-          return require('path').isAbsolute(p) ? p : require('path').join(cwd, p);
-        }
-        function typeToSeverity(type) {
-          switch (type && type.toLowerCase()) {
-            case 'err':
-            case 'error': return 'error';
-            case 'warn':
-            case 'warning': return 'warning';
-            default: return null;
-          }
-        }
-
-        this.linter && this.linter.setMessages(this.errorMatcher.getMatches().map(match => ({
-          type: match.type || 'Error',
-          text: match.message || 'Error from build',
-          filePath: normalizePath(match.file),
-          severity: typeToSeverity(match.type),
-          range: extractRange(match),
-          trace: match.trace && match.trace.map(trace => ({
-            type: trace.type || 'Trace',
-            text: trace.message || 'Trace in build',
-            filePath: trace.file && normalizePath(trace.file),
-            severity: typeToSeverity(trace.type) || 'info',
-            range: extractRange(trace)
-          }))
-        })));
+        this.linter && this.linter.processMessages(this.errorMatcher.getMatches(), cwd);
 
         if (atom.config.get('build.beepWhenDone')) {
           atom.beep();
@@ -324,7 +292,9 @@ export default {
   },
 
   consumeLinterRegistry(registry) {
-    this.linter = registry.register({ name: 'Build' });
+    this.linter && this.linter.destroy();
+    const Linter = require('linter-integration');
+    this.linter = new Linter(registry);
   },
 
   consumeBuilder(builder) {
